ints to double, in case of temps with comma places

+++ Week7Converters/Program.cs
@@ -25,16 +25,16 @@ namespace Week7Converters
         public static void CelsiusFahrenheit()
         {
             Console.WriteLine("Enter your temperature in celsius.");
-            int userInput = Int32.Parse(Console.ReadLine());
-            int fahrenheitResult = (userInput * 9) / 5 + 32;
+            double userInput = Double.Parse(Console.ReadLine());
+            double fahrenheitResult = (userInput * 9) / 5 + 32;
             Console.WriteLine($"{userInput} celsius is {fahrenheitResult} fahrenheit.");
 
         }
         public static void FahrenheitCelsius()
         {
             Console.WriteLine("Enter your temperature in fahrenheit.");
-            int userInput = Int32.Parse(Console.ReadLine());
-            int celsiusResult = (userInput - 32) * 5 / 9;
+            double userInput = Double.Parse(Console.ReadLine());
+            double celsiusResult = (userInput - 32) * 5 / 9;
             Console.WriteLine($"{userInput} in fahrenheit is {celsiusResult} celsius.");
         }
     }
